changed properties to type Double

+++ mainproject/src/main/scala/com/raphtory/testCases/wordSemantic/graphbuilders/CSVGraphBuilder.scala
@@ -20,7 +20,7 @@ class CSVGraphBuilder extends GraphBuilder[String] {
             updateTime = occurenceTime,
             srcId = srcClusterId,
             dstId = dstClusterId,
-            Properties(LongProperty("Frequency", coocWeight))
+            Properties(DoubleProperty("Frequency", coocWeight))
           )
 
     } catch {
